feat(worldcore): 添加获取实体所有组件方法及相关优化
- 在Chunk类中新增GetAll方法，用于获取指定实体索引的所有组件对象
- 在Chunk类中新增GetAllShared方法，获取所有共享组件对象
- 在WorldCore.ComponentSimple中添加Get和GetAll方法，支持根据实体和组件类型获取单个或所有组件
- 在WorldCore.ComponentStatic中添加GetSta方法，实现静态组件获取功能
- 修改EntityInfo中ArchetypeIndex字段类型为ushort，提升类型精度
- 修正SharedComponentGroup的GetIndexFromHash方法，避免抛出异常，返回索引或-1
- 在WorldCore.Entity中添加HasEntityAndNotDestroyed方法，用于判断实体是否存在且未销毁
- 在ComponentInfo中增加隐式转换操作符，简化ComponentMeta访问
- 在单元测试中增加对获取所有组件方法的测试，验证功能正确性

+++ FLib/Sources/WorldCores/Entities/EntityInfo.cs
@@ -20,7 +20,7 @@ namespace FLib.WorldCores
         /// <summary>
         /// 
         /// </summary>
-        public readonly int ArchetypeIndex => _destroyFlagWithArchetypeIndex & 0x7fff;
+        public readonly ushort ArchetypeIndex => (ushort)(_destroyFlagWithArchetypeIndex & 0x7fff);
 
         /// <summary>
         ///
